Added new DI modules.
